+++ core/output/formatter.go
@@ -1,32 +1,85 @@
 package output
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	"io"
+	"sort"
+	"strings"
+	"text/tabwriter"
 	plainTemplate "text/template"
 
-	color "github.com/gookit/color"
+	fColor "github.com/fatih/color"
+	gColor "github.com/gookit/color"
+	"github.com/nathan-fiscaletti/consolesize-go"
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 
 	"github.com/reliablyhq/cli/core"
+	"github.com/reliablyhq/cli/core/color"
+	"github.com/reliablyhq/cli/utils"
 	"github.com/reliablyhq/cli/version"
 )
 
-var text = `Results: {{ range $index, $issue := .Suggestions }}
-[{{ highlight $issue.FileLocation 0 }}] - {{ $issue.RuleID }} : {{ $issue.Message }} (Platform: {{ $issue.Platform}}, Kind: {{ $issue.Kind }}){{ end }}
+const sortFlag = true
+
+var text = `{{ notice "Results:" }} {{ range $index, $issue := .Suggestions }}
+{{ prompt ">" }} {{ $issue.FileLocation }} [{{ printLevel $issue.Level }}] {{ $issue.Message }}
+Rule: {{ $issue.RuleID }}, Platform: {{ $issue.Platform}}, Kind: {{ $issue.Kind }}
+{{if $issue.Example }}
+# Example:
+{{ $issue.Example}}
+{{ end }}{{ end }}
 {{ notice "Summary:" }}
-	{{ $count := len .Suggestions }}Suggestions: {{ if eq $count 0 }}
+	{{ $count := len .Suggestions }}{{ if eq $count 0 }}
 	{{- success "No suggestion found" }}
-	{{- else }}
-	{{- danger $count " suggestion(s) found" }}
+	{{- else }}{{ if eq $count 1 }}
+			{{- danger $count " suggestion found" }}
+		{{- else }}
+			{{- danger $count " suggestions found" }}
+		{{- end }}
+	{{ counter "info" .Counters.info }} - {{ counter "warning" .Counters.warning }} - {{ counter "error" .Counters.error }}
 	{{- end }}
 
 `
 
+var textTemplateTabbed = `Results:{{ range $index, $issue := .Suggestions }}
+{{ levelSymbol $issue.Level}}	{{ printLiveFileLocation $issue.FileLocation $issue.Kind }}	{{ $issue.Platform}}:{{ $issue.Kind }}	{{ $issue.RuleID  }}	{{ $issue.Message }} {{ end }}
+{{ notice "Summary:" }}
+	{{ $count := len .Suggestions }}{{ if eq $count 0 }}
+	{{- success "No suggestion found" }}
+	{{- else }}{{ if eq $count 1 }}
+			{{- danger $count " suggestion found" }}
+		{{- else }}
+			{{- danger $count " suggestions found" }}
+		{{- end }}
+	{{ counter "info" .Counters.info }} - {{ counter "warning" .Counters.warning }} - {{ counter "error" .Counters.error }}
+	{{- end }}
+
+	`
+
 type reportInfo struct {
 	Suggestions []*core.Suggestion `json:"suggestions"`
+	Counters    map[string]int
+}
+
+type bySuggestion []*core.Suggestion
+
+func countsPerLevel(suggestions []*core.Suggestion) map[string]int {
+
+	var counters map[string]int = map[string]int{
+		"info":    0,
+		"warning": 0,
+		"error":   0,
+	}
+
+	for _, s := range suggestions {
+		l := s.Level.String()
+		counters[l] = counters[l] + 1
+	}
+
+	return counters
 }
 
 // CreateReport generates a report for the reported suggestions into
@@ -42,6 +95,7 @@ func CreateReport(w io.Writer, format string, baseDir string, suggestions []*cor
 
 	data := &reportInfo{
 		Suggestions: suggestions,
+		Counters:    countsPerLevel(suggestions),
 	}
 	var err error
 	switch format {
@@ -49,8 +103,10 @@ func CreateReport(w io.Writer, format string, baseDir string, suggestions []*cor
 		err = reportJSON(w, data)
 	case "yaml":
 		err = reportYAML(w, data)
-	case "text":
+	case "text", "extended":
 		err = reportFromPlaintextTemplate(w, text, data)
+	case "tabbed":
+		err = reportFromTabbedTextTemplate(w, data)
 	case "simple", "basic", "linter":
 		err = reportLinter(w, data)
 	case "sarif":
@@ -101,7 +157,7 @@ func reportLinter(w io.Writer, data *reportInfo) error {
 		*/
 		_, err := fmt.Fprintf(w, "%s [%s] %s\n",
 			s.FileLocation(),
-			s.Level,
+			s.Level.ColoredString(),
 			s.Message,
 		)
 		if err != nil {
@@ -240,7 +296,7 @@ func levelToSarifLevel(l core.Level) sarifLevel {
 }
 
 func reportFromPlaintextTemplate(w io.Writer, reportTemplate string, data *reportInfo) error {
-	enableColor := true
+	enableColor := !fColor.NoColor
 	t, e := plainTemplate.
 		New("reliably").
 		Funcs(plainTextFuncMap(enableColor)).
@@ -252,14 +308,86 @@ func reportFromPlaintextTemplate(w io.Writer, reportTemplate string, data *repor
 	return t.Execute(w, data)
 }
 
+func (s bySuggestion) Len() int {
+	return len(s)
+}
+func (s bySuggestion) Swap(i, j int) {
+	s[i], s[j] = s[j], s[i]
+}
+func (s bySuggestion) Less(i, j int) bool {
+	return s[i].Level > s[j].Level
+}
+
+func reportFromTabbedTextTemplate(w io.Writer, data *reportInfo) error {
+
+	padding := 2
+	minWidth := 0
+	tabWidth := 0
+	const padChar = ' '
+
+	reportBuffer := new(bytes.Buffer)
+
+	tw := tabwriter.NewWriter(reportBuffer, minWidth, tabWidth, padding, padChar, 0)
+	if sortFlag {
+		sort.Sort(bySuggestion(data.Suggestions))
+	}
+
+	consoleColumns, _ := consolesize.GetConsoleSize()
+
+	err := reportFromPlaintextTemplate(tw, textTemplateTabbed, data)
+
+	if err != nil {
+
+		return err
+	}
+	tw.Flush()
+
+	reportString := reportBuffer.String()
+	reportLines := strings.Split(reportString, "\n")
+	for _, reportLine := range reportLines {
+
+		// the truncate is off by 3 characters because of the ■ character at the
+		// begining of the line
+		fmt.Println(utils.TruncateString(reportLine, consoleColumns))
+
+	}
+
+	return nil
+}
+
 func plainTextFuncMap(enableColor bool) plainTemplate.FuncMap {
 	if enableColor {
 		return plainTemplate.FuncMap{
 			"highlight": highlight,
-			"danger":    color.Danger.Render,
-			"notice":    color.Notice.Render,
-			"success":   color.Success.Render,
+			"danger":    gColor.Danger.Render,
+			"notice":    gColor.Notice.Render,
+			"success":   gColor.Success.Render,
+			"prompt":    gColor.Yellow.Render,
 			"printCode": fmt.Sprint,
+			"printLevel": func(l core.Level) string {
+				return l.ColoredString()
+			},
+			"printLiveFileLocation": func(s string, kind string) string {
+				r := kind + ":"
+				return strings.Replace(s, r, "", 1)
+			},
+			"levelSymbol": func(l core.Level) string {
+				return l.ColoredSquare()
+			},
+			"counter": func(level string, count int) string {
+				var tick string
+				switch level {
+				case "info":
+					tick = color.BgYellow(" ")
+				case "warning":
+					tick = color.BgMagenta(" ")
+				case "error":
+					tick = color.BgRed(" ")
+				default:
+					tick = ""
+				}
+				return fmt.Sprintf("%s %v %s", tick, count, level)
+			},
 		}
 	}
 
@@ -271,14 +399,28 @@ func plainTextFuncMap(enableColor bool) plainTemplate.FuncMap {
 		"danger":    fmt.Sprint,
 		"notice":    fmt.Sprint,
 		"success":   fmt.Sprint,
+		"prompt":    fmt.Sprint,
 		"printCode": fmt.Sprint,
+		"printLevel": func(l core.Level) string {
+			return l.String()
+		},
+		"printLiveFileLocation": func(s string, kind string) string {
+			r := kind + ":"
+			return strings.Replace(s, r, "", 1)
+		},
+		"levelSymbol": func(l core.Level) string {
+			return l.String()[0:1]
+		},
+		"counter": func(level string, count int) string {
+			return fmt.Sprintf("%v %s", count, level)
+		},
 	}
 }
 
 var (
-	errorTheme   = color.New(color.FgLightWhite, color.BgRed)
-	warningTheme = color.New(color.FgBlack, color.BgYellow)
-	defaultTheme = color.New(color.FgWhite, color.BgBlack)
+	errorTheme   = gColor.New(gColor.FgLightWhite, gColor.BgRed)
+	warningTheme = gColor.New(gColor.FgBlack, gColor.BgYellow)
+	defaultTheme = gColor.New(gColor.FgWhite, gColor.BgBlack)
 )
 
 // highlight returns content t colored based on Score